Refactor protobuf model
