+++ persistence/src/main/protobuf/filing.proto
@@ -1,0 +1,16 @@
+syntax = "proto3";
+
+package hmda.model;
+
+enum FilingStatusMessage {
+    NOT_STARTED = 0;
+    IN_PROGRESS = 1;
+    COMPLETED = 2;
+    CANCELLED = 3;
+}
+
+message FilingMessage {
+    string period = 1;
+    string institutionId = 2;
+    FilingStatusMessage status = 3;
+}
