+++ persistence/src/main/scala/hmda/persistence/institutions/serialization/SubmissionConverter.scala
@@ -1,0 +1,6 @@
+package hmda.persistence.institutions.serialization
+
+object SubmissionConverter {
+
+  //implicit def messageToSubmission(m: Option[])
+}
